[Oztechan/CCC#1757] Create android:feature:widget module for widget
