+++ android/feature/widget/src/main/kotlin/com/oztechan/ccc/android/feature/widget/ui/components/ImageView.kt
@@ -1,0 +1,16 @@
+package com.oztechan.ccc.android.feature.widget.ui.components
+
+import androidx.compose.runtime.Composable
+import androidx.glance.GlanceModifier
+import androidx.glance.Image
+import androidx.glance.ImageProvider
+
+@Composable
+fun ImageView(
+    provider: ImageProvider,
+    modifier: GlanceModifier = GlanceModifier
+) = Image(
+    provider = provider,
+    contentDescription = "",
+    modifier = modifier
+)
