+++ android/feature/widget/src/main/res/drawable/ic_next.xml
