+++ android/feature/widget/src/main/res/drawable/ic_sync_widget.xml
