+++ android/feature/widget/src/main/res/xml/app_widget_info.xml
