--- android/widget/build.gradle.kts
@@ -1,43 +1,0 @@
-plugins {
-    @Suppress("DSL_SCOPE_VIOLATION")
-    libs.plugins.apply {
-        id(androidLib.get().pluginId)
-        id(android.get().pluginId)
-    }
-}
-
-@Suppress("UnstableApiUsage")
-android {
-    ProjectSettings.apply {
-        namespace = Modules.Android.widget.packageName
-        compileSdk = COMPILE_SDK_VERSION
-
-        defaultConfig {
-            minSdk = MIN_SDK_VERSION
-            targetSdk = TARGET_SDK_VERSION
-        }
-    }
-
-    buildFeatures {
-        compose = true
-    }
-
-    composeOptions {
-        kotlinCompilerExtensionVersion = libs.versions.compose.get()
-    }
-}
-
-dependencies {
-    libs.android.apply {
-        implementation(glance)
-        implementation(koinAndroid)
-    }
-
-    Modules.apply {
-        implementation(project(client))
-        implementation(project(res))
-        implementation(project(test))
-    }
-
-    implementation(project(Modules.Submodules.logmob))
-}
